requirements.txt
aiogram==3.13.1
pybit==5.7.0
pandas==2.2.2
pyarrow==16.1.0
numpy==1.26.4
python-dateutil==2.9.0.post0
tomli==2.0.1
pydantic==2.8.2
APScheduler==3.10.4
python-dotenv==1.0.1
pycryptodome==3.20.0

